Add interceptors

+++ Backend/Projet2024/Domaine/CourseNote.cs
@@ -18,7 +18,7 @@ namespace Domaine
         public int UserId { get; set; }
         public User User { get; set; }
 
-        public int Note { get; set; }
+        public float Note { get; set; }
 
 
     }
